pull

+++ includes/hostapd.php
@@ -1,6 +1,8 @@
 <?php
 
 include_once('includes/status_messages.php');
+include_once('app/lib/system.php');
+require_once 'config.php';
 
 /**
 *
@@ -9,6 +11,7 @@ include_once('includes/status_messages.php');
 function DisplayHostAPDConfig()
 {
     $status = new StatusMessages();
+	$system = new System();
     $arrHostapdConf = parse_ini_file('/etc/raspap/hostapd.ini');
     $arrConfig = array();
     $arr80211Standard = [
@@ -23,33 +26,34 @@ function DisplayHostAPDConfig()
     $managedModeEnabled = false;
     exec("ip -o link show | awk -F': ' '{print $2}'", $interfaces);
 
-    if (isset($_POST['SaveHostAPDSettings'])) {
-        SaveHostAPDConfig($arrSecurity, $arrEncType, $arr80211Standard, $interfaces, $status);
-    } elseif (isset($_POST['StartHotspot'])) {
-        $status->addMessage('Attempting to start hotspot', 'info');
-        if ($arrHostapdConf['WifiAPEnable'] == 1) {
-            exec('sudo /etc/raspap/hostapd/servicestart.sh --interface uap0 --seconds 3', $return);
-        } else {
-            exec('sudo /etc/raspap/hostapd/servicestart.sh --seconds 5', $return);
-        }
-        foreach ($return as $line) {
-            $status->addMessage($line, 'info');
-        }
-    } elseif (isset($_POST['StopHotspot'])) {
-        $status->addMessage('Attempting to stop hotspot', 'info');
-        exec('sudo /bin/systemctl stop hostapd.service', $return);
-        foreach ($return as $line) {
-            $status->addMessage($line, 'info');
+    if (!RASPI_MONITOR_ENABLED) {
+        if (isset($_POST['SaveHostAPDSettings'])) {
+            SaveHostAPDConfig($arrSecurity, $arrEncType, $arr80211Standard, $interfaces, $status);
+        } elseif (isset($_POST['StartHotspot'])) {
+            $status->addMessage('Attempting to start hotspot', 'info');
+            if ($arrHostapdConf['WifiAPEnable'] == 1) {
+                exec('sudo /etc/raspap/hostapd/servicestart.sh --interface uap0 --seconds 3', $return);
+            } else {
+                exec('sudo /etc/raspap/hostapd/servicestart.sh --seconds 5', $return);
+            }
+            foreach ($return as $line) {
+                $status->addMessage($line, 'info');
+            }
+        } elseif (isset($_POST['StopHotspot'])) {
+            $status->addMessage('Attempting to stop hotspot', 'info');
+            exec('sudo /bin/systemctl stop hostapd.service', $return);
+            foreach ($return as $line) {
+                $status->addMessage($line, 'info');
+            }
         }
     }
 
     exec('cat '. RASPI_HOSTAPD_CONFIG, $hostapdconfig);
-    exec('pidof hostapd | wc -l', $hostapdstatus);
     exec('iwgetid '. RASPI_WIFI_CLIENT_INTERFACE. ' -r', $wifiNetworkID);
     if ( !empty($wifiNetworkID[0])) {
         $managedModeEnabled = true;
     }
-
+    $hostapdstatus = $system->hostapdStatus();
     $serviceStatus = $hostapdstatus[0] == 0 ? "down" : "up";
 
     foreach ($hostapdconfig as $hostapdconfigline) {
